refactor, fix(image, progress, lqip): review polish round 15
Significant:
- lqip-encoder split into two modules. lqip-encoder.ts is now pure —
  no Lit imports, no @customElement side effects — so a Node consumer
  importing encodeLqip / encodePixelDataToLqip no longer drags
  nldd-lqip-encoder into their custom-element registry. The Storybook
  UI element moves to lqip-encoder-element.ts and registers itself
  there. Storybook story updated to import the element module instead.
- nldd-progress: replace the undiscoverable text=" " (space-as-sentinel)
  with an explicit `no-label` boolean attribute. The space hack only
  lived in a TODO comment; consumers reading the @attr docs would never
  have found it. New story "Without label" demonstrates the API.

Minor:
- nldd-image: height now uses reflect: true to match width. Both are
  intrinsic-dimension hints; one reflecting and the other not made the
  HTML serialisation inconsistent.
- progress-bar.test: replace expect(true).toBe(true) tautology with
  expect(() => cleanup(el)).not.toThrow(), which actually expresses
  what the test is verifying (clean disconnect path).
- progress-bar.ts: stale comment referenced _renderHeaderValue
  (renamed long ago) — now points at the _displayValue getter that
  actually performs the strip.
- lqip-encoder.ts: bitsToLab JSDoc previously claimed wire integers
  are "interchangeable with Lean's tool". Per-byte L/a/b decoding does
  match, but the overall CSV protocol does not — clarify the boundary
  so a reader of just this function doesn't get the wrong impression.
